requirements.txt
fastapi
uvicorn
langchain_groq
langchain_core
python-dotenv